+++ features/Context/Page/Base/Form.php
@@ -344,6 +344,15 @@ class Form extends Base
         return $this->getElement('Updates grid')->findAll('css', 'tbody tr');
     }
 
+    public function expandAttribute($attribute)
+    {
+        if (null === $label = $this->find('css', sprintf('label:contains("%s")', $attribute))) {
+            throw new \InvalidArgumentException(sprintf('Cannot find attribute "%s" field', $attribute));
+        }
+
+        return $this->expand($label);
+    }
+
     /**
      * @param string $label
      */