+++ src/Pim/Bundle/CatalogBundle/Entity/Attribute.php
@@ -35,6 +35,9 @@ class Attribute extends AbstractEntityAttribute implements
      */
     protected $options;
 
+    /**
+     * @var integer $sortOrder
+     */
     protected $sortOrder = 0;
 
     /**
@@ -151,17 +154,40 @@ class Attribute extends AbstractEntityAttribute implements
      */
     protected $translations;
 
+    /**
+     * @var boolean $required
+     */
     protected $required;
 
+    /**
+     * @var boolean $unique
+     */
     protected $unique;
 
+    /**
+     * @var mixed $defaultValue
+     */
     protected $defaultValue;
 
+    /**
+     * @var boolean $searchable
+     */
     protected $searchable;
 
+    /**
+     * @var boolean $localizable
+     */
     protected $localizable;
 
+    /**
+     * @var boolean $scopable
+     */
     protected $scopable;
+
+    /**
+     * @var array
+     */
+    protected $properties;
 
     /**
      * Constructor
@@ -181,6 +207,7 @@ class Attribute extends AbstractEntityAttribute implements
         $this->families            = new ArrayCollection();
         $this->translations        = new ArrayCollection();
         $this->validationRule      = null;
+        $this->properties          = array();
     }
 
     /**
@@ -1003,5 +1030,19 @@ class Attribute extends AbstractEntityAttribute implements
     public function getReference()
     {
         return $this->code;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setAttributeType($type)
+    {
+        parent::setAttributeType($type);
+
+        if ($this->attributeType === 'pim_catalog_identifier') {
+            $this->required = true;
+        }
+
+        return $this;
     }
 }