+++ src/Pim/Bundle/DataGridBundle/Extension/Pager/MongoDbOdm/Pager.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace Pim\Bundle\DataGridBundle\Extension\Pager\MongoDbOdm;
+
+use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
+use Oro\Bundle\DataGridBundle\Extension\Pager\PagerInterface;
+use Oro\Bundle\DataGridBundle\Extension\Pager\AbstractPager;
+use Oro\Bundle\SecurityBundle\ORM\Walker\AclHelper;
+
+/**
+ * MongoDB ODM pager
+ *
+ * @author    Nicolas Dupont <nicolas@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class Pager extends AbstractPager implements PagerInterface
+{
+    /**
+     * @var QueryBuilder
+     */
+    protected $qb;
+
+    /**
+     * @var AclHelper
+     */
+    protected $aclHelper;
+
+    public function __construct(AclHelper $aclHelper, $maxPerPage = 10, QueryBuilder $qb = null)
+    {
+        $this->qb = $qb;
+        parent::__construct($maxPerPage);
+        $this->aclHelper = $aclHelper;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function init()
+    {
+        throw new \RuntimeException("Not implemented yet ! ".__CLASS__."::".__METHOD__);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getResults($hydrationMode = Query::HYDRATE_OBJECT)
+    {
+        throw new \RuntimeException("Not implemented yet ! ".__CLASS__."::".__METHOD__);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function retrieveObject($offset)
+    {
+        throw new \RuntimeException("Not implemented yet ! ".__CLASS__."::".__METHOD__);
+    }
+
+    /**
+     * @param QueryBuilder $qb
+     *
+     * @return $this
+     */
+    public function setQueryBuilder(QueryBuilder $qb)
+    {
+        $this->qb = $qb;
+
+        return $this;
+    }
+
+    /**
+     * @return QueryBuilder
+     */
+    public function getQueryBuilder()
+    {
+        return $this->qb;
+    }
+}