+++ src/Pim/Bundle/DataGridBundle/Extension/Pager/OrmPagerExtension.php
@@ -6,7 +6,6 @@ use Oro\Bundle\DataGridBundle\Datagrid\Common\DatagridConfiguration;
 use Oro\Bundle\DataGridBundle\Datagrid\Builder;
 use Oro\Bundle\DataGridBundle\Extension\Pager\OrmPagerExtension as OroOrmPagerExtension;
 use Pim\Bundle\DataGridBundle\Datasource\Orm\OrmDatasource as PimOrmDatasource;
-use Pim\Bundle\DataGridBundle\Datasource\ProductDatasource;
 
 /**
  * Orm pager extension
@@ -34,7 +33,7 @@ class OrmPagerExtension extends OroOrmPagerExtension
     {
         $datasourceType = $config->offsetGetByPath(Builder::DATASOURCE_TYPE_PATH);
 
-        return ($datasourceType == PimOrmDatasource::TYPE || $datasourceType == ProductDatasource::TYPE);
+        return ($datasourceType == PimOrmDatasource::TYPE);
     }
 
     /**