+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.1.0 -
+
+## Features
+
+## Improvements
+- Removed useless ```app/entities``` directory
+- Add a 'properties' field to the Attribute entity to allow easily adding new attribute type dependent properties
+
+## Bug fixes
+- Fixed verbose option always used
+- Fixed issue on attribute option removal
+- Fixed identifier is required attribute
+
+## BC breaks
+- Add an argument HydratorInterface in ProductDatasource constructor (MongoDBODM support)
+- Remove deprecated fallback property of Locale entity
+
 # 1.0.0 - "Hare We Go" (2014-03-06)
 
 ## Features